refactor(coordination): extract ceremony policy resolver out of controller
Move the ceremony-policy read helpers (settings parsing, project /
department fetch, resolution + per-field origin tracking) and the
ResolvedCeremonyPolicyResponse / ActiveCeremonyStrategyResponse
response models out of api/controllers/ceremony_policy.py into a new
coordination-domain module: coordination/ceremony_policy/policy_resolver.py.

The MCP service layer (CeremonyPolicyService) now imports those
helpers directly from the resolver instead of doing lazy imports out
of the api/ controller, breaking the controller versus service
circular dependency that the dependency-inversion audit flagged.

Tests now patch the bound names inside the service module rather
than reaching into the controller.

Part of #1773 (Phase 2).
